pyproject.toml
[tool.poetry]
name = "bot-task-scheduler"
version = "0.1.0"
description = ""
authors = ["messiah1349 <messiah1349@gmail.com>"]
readme = "README.md"
packages = [{include = "bot_task_scheduler"}]

[tool.poetry.dependencies]
python = "^3.10"
anyio = "3.7.1"
blinker = "1.6.2"
certifi = "2023.5.7"
charset-normalizer = "3.2.0"
click = "8.1.6"
exceptiongroup = "1.1.2"
flask = "2.3.2"
h11 = "0.14.0"
httpcore = "0.16.3"
httpx = "0.23.3"
idna = "3.4"
itsdangerous = "2.1.2"
jinja2 = "3.1.2"
markupsafe = "2.1.3"
marshmallow = "3.20.1"
packaging = "23.1"
python-telegram-bot = "20.0a6"
requests = "2.31.0"
rfc3986 = "1.5.0"
scheduler = "0.8.4"
sniffio = "1.3.0"
typeguard = "4.0.0"
typing-extensions = "4.7.1"
urllib3 = "2.0.3"
webargs = "8.3.0"
werkzeug = "2.3.6"
apscheduler = "^3.10.1"
python-dateutil = "^2.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
